requirements.txt
blinker==1.8.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dnspython==2.6.1
dsinternals==1.2.4
Flask==3.0.3
future==1.0.0
idna==3.7
impacket==0.11.0
itsdangerous==2.2.0
Jinja2==3.1.4
ldap3==2.9.1
ldapdomaindump==0.9.4
MarkupSafe==2.1.5
pyasn1==0.6.0
pycparser==2.22
pycryptodomex==3.20.0
pyOpenSSL==24.1.0
pyspnego==0.11.0
pywinrm==0.4.3
requests==2.32.3
requests-credssp==2.0.0
requests_ntlm==1.3.0
six==1.16.0
tabulate==0.9.0
tqdm==4.66.4
urllib3==2.2.1
Werkzeug==3.0.3
xmltodict==0.13.0
